pyproject.toml
[project]
name = "ams_print"
version = "0.1.0"
dependencies = [
  "Pillow==10.3.0",
  "typer==0.12.3"
]

[project.scripts]
ams_print = "ams_print.cli:app"
